[metadata/genius] improved lyrics fetcher
Genius’ search is limited to ten results and not able to paginate.
Also, it searches everywhere and therefore most likely returns songs.
That’s why it is better to search for artist and song title first to
find an artist or a song.
The old approach to search for the artist only is now only used as fallback.